Remove diffing for initial implementation and add err.sh
