+++ packages/next/package.json
@@ -68,7 +68,6 @@
     "@babel/runtime": "7.4.5",
     "@babel/runtime-corejs2": "7.4.5",
     "amphtml-validator": "1.0.23",
-    "ansi-to-html": "0.6.11",
     "async-sema": "3.0.0",
     "autodll-webpack-plugin": "0.4.2",
     "babel-core": "7.0.0-bridge.0",
@@ -87,7 +86,6 @@
     "next-server": "9.0.4",
     "prop-types": "15.7.2",
     "prop-types-exact": "1.2.0",
-    "raw-body": "2.4.1",
     "react-error-overlay": "6.0.1",
     "react-is": "16.8.6",
     "serialize-javascript": "1.7.0",
